cataclysm: fix build on Linux (#99597)
* cataclysm: fix build on Linux

* Update Formula/cataclysm.rb


* cataclysm: kill process in test only if pid is defined

+++ Formula/cataclysm.rb
@@ -31,10 +31,10 @@ class Cataclysm < Formula
   depends_on "sdl2_ttf"
 
   def install
+    os = OS.mac? ? "osx" : OS.kernel_name.downcase
     args = %W[
-      NATIVE=osx
+      NATIVE=#{os}
       RELEASE=1
-      OSX_MIN=#{MacOS.version}
       USE_HOME_DIR=1
       TILES=1
       SOUND=1
@@ -43,6 +43,7 @@ class Cataclysm < Formula
       LINTJSON=0
     ]
 
+    args << "OSX_MIN=#{MacOS.version}" if OS.mac?
     args << "CLANG=1" if ENV.compiler == :clang
 
     system "make", *args
@@ -57,18 +58,24 @@ class Cataclysm < Formula
   end
 
   test do
+    # Disable test on Linux because it fails with this error:
+    # Error while initializing the interface: SDL_Init failed: No available video device
+    return if OS.linux? && ENV["HOMEBREW_GITHUB_ACTIONS"]
+
     # make user config directory
     user_config_dir = testpath/"Library/Application Support/Cataclysm/"
     user_config_dir.mkpath
 
-    # run cataclysm for 7 seconds
+    # run cataclysm for 30 seconds
     pid = fork do
       exec bin/"cataclysm"
     end
-    sleep 30
-    assert_predicate user_config_dir/"config",
-                     :exist?, "User config directory should exist"
-  ensure
-    Process.kill("TERM", pid)
+    begin
+      sleep 30
+      assert_predicate user_config_dir/"config",
+                       :exist?, "User config directory should exist"
+    ensure
+      Process.kill("TERM", pid)
+    end
   end
 end
